+++ tests/test_manager.py
@@ -1,10 +1,73 @@
 import copy
 import numpy as np
+import paho.mqtt.client as mqtt
 from unittest import TestCase
 from NoveltyProducer.Manager import Manager, InvalidInputTypeError, InvalidInputValueError
+from NoveltyProducer.Generator import Generator
+from apscheduler.job import Job
+
+# ext_ip = 'test.mosquitto.org'
+ip = 'localhost' 
+port = 1883
+topic = 'foo'
+frequency = 1
+channel_name = 'bar'
+channel_limits = [-2, 2]
+channel_frequency = 0.1
+novelty_frequency = 0.01
+novelty_duration = 1
+novelty_impact = 1
+pipeline_name = 'pipe'
 
 class TestBaseUnit(TestCase):
 
+    def test_value_output(self):
+        """Test output of each function."""
+        
+        # init instance of Manager
+        man = Manager()
+        
+        # _add_connection
+        con_id = man._add_connection(ip_=ip, port_=port)
+        self.assertIn(ip, man.connections[con_id]['ip'])
+        self.assertIn(port, man.connections[con_id]['port'])
+        
+        # _add_topic
+        top_id = man._add_topic(topic_=topic, frequency_=frequency)
+        self.assertIn(topic, man.topics[top_id]['topic'])
+        self.assertIn(frequency, man.topics[top_id]['frequency'])
+        
+        # _add_channel
+        chn_id = man._add_channel(name_=channel_name, limits_=channel_limits, frequency_=channel_frequency)
+        self.assertIn(channel_name, man.channels[chn_id]['name'])
+        self.assertIn(channel_limits, man.channels[chn_id]['limits'])
+        self.assertIn(channel_frequency, man.channels[chn_id]['frequency'])
+        
+        # _add_novelty
+        nov_id = man._add_novelty(frequency_=novelty_frequency, duration_=novelty_duration, impact_=novelty_impact)
+        self.assertIn(novelty_frequency, man.novelties[nov_id]['frequency'])
+        self.assertIn(novelty_duration, man.novelties[nov_id]['duration'])
+        self.assertIn(novelty_impact, man.novelties[nov_id]['impact'])
+        
+        # _add_pipeline
+        pipe_id = man._add_pipeline(name_=pipeline_name, host_id_=con_id, topic_id_=top_id, channel_id_=chn_id, novelty_id_=nov_id)
+        self.assertIn(pipeline_name, man.pipelines[pipe_id]['name'])
+        self.assertIn(con_id, man.pipelines[pipe_id]['host_id'])
+        self.assertIn(top_id, man.pipelines[pipe_id]['topic_id'])
+        self.assertIn(chn_id, man.pipelines[pipe_id]['channel_id'])
+        self.assertIn(nov_id, man.pipelines[pipe_id]['novelty_id'])
+        self.assertIn(1, man.pipelines[pipe_id]['active'])
+        
+        # switch_pipeline (on to off)
+        pipeline_status = man.pipelines[pipe_id]['active']
+        man.switch_pipeline(pipe_id)
+        self.assertTrue(pipeline_status != man.pipelines[pipe_id]['active'])
+        
+        # switch_pipeline (off to on)
+        pipeline_status = man.pipelines[pipe_id]['active']
+        man.switch_pipeline(pipe_id)
+        self.assertTrue(pipeline_status != man.pipelines[pipe_id]['active'])
+        
     def test_type_output(self):
         """Test types of output"""
         
@@ -12,24 +75,34 @@ class TestBaseUnit(TestCase):
         man = Manager()
         
         # _add_connection
-        con_id = man._add_connection(ip_='localhost', port_=1883)
-        self.assertTrue(isinstance(con_id, int)) 
+        con_id = man._add_connection(ip_=ip, port_=port)
+        self.assertIsInstance(con_id, int)
         
         # _add_topic
-        top_id = man._add_topic(topic_='foo', frequency_=1)
-        self.assertTrue(isinstance(top_id, int)) 
+        top_id = man._add_topic(topic_=topic, frequency_=frequency)
+        self.assertIsInstance(top_id, int)
         
         # _add_channel
-        chn_id = man._add_channel(name_='bar', limits_=[-1, 1], frequency_=0.1)
-        self.assertTrue(isinstance(chn_id, int))
+        chn_id = man._add_channel(name_=channel_name, limits_=channel_limits, frequency_=channel_frequency)
+        self.assertIsInstance(chn_id, int)
         
         # _add_novelty
-        nov_id = man._add_novelty(frequency_=0.01, duration_=1, impact_=1)
-        self.assertTrue(isinstance(nov_id, int))
+        nov_id = man._add_novelty(frequency_=novelty_frequency, duration_=novelty_duration, impact_=novelty_impact)
+        self.assertIsInstance(nov_id, int)
         
         # _add_pipeline
         pipe_id = man._add_pipeline(name_='pipe', host_id_=con_id, topic_id_=top_id, channel_id_=chn_id, novelty_id_=nov_id)
-        self.assertTrue(isinstance(pipe_id, int))
+        self.assertIsInstance(pipe_id, int)
+        
+        # _add_handlers
+        man._add_handlers(pipe_id)
+        self.assertIsInstance(man.handlers[pipe_id]['generator'], Generator)
+        self.assertIsInstance(man.handlers[pipe_id]['mqtt'], mqtt.Client)
+        
+        # create_pipeline
+        man.create_pipeline(ip_=ip, port_=port, topic_=topic, channel_name_=channel_name, pipeline_name_='sffresch')
+        pipe_id = str([k for k,v in man.pipelines.items() if man.pipelines[k]['name'] == 'sffresch'][0])
+        self.assertIsInstance(man.Scheduler.get_job(pipe_id), Job)
 
     def test_invalid_inputs(self):
         """Test for all expected errors that should be raised when given invalid inputs"""
@@ -39,58 +112,58 @@ class TestBaseUnit(TestCase):
         
         # ip_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_=127001, port_=1883, topic_='foo', frequency_=1, channel_name_='bar')
+            invalid_generator = man.create_pipeline(ip_=127001, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name)
         
         # port_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_='1883', topic_='foo', frequency_=1, channel_name_='bar')
+            invalid_generator = man.create_pipeline(ip_=ip, port_='1883', topic_=topic, frequency_=frequency, channel_name_=channel_name)
         with self.assertRaises(InvalidInputValueError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=65536, topic_='foo', frequency_=1, channel_name_='bar')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=65536, topic_=topic, frequency_=frequency, channel_name_=channel_name)
         with self.assertRaises(InvalidInputValueError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=-1, topic_='foo', frequency_=1, channel_name_='bar')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=-1, topic_=topic, frequency_=frequency, channel_name_=channel_name)
             
         # topic_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_=42, frequency_=1, channel_name_='bar')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=42, frequency_=frequency, channel_name_=channel_name)
 
         # frequency_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_='1', channel_name_='bar')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_='1', channel_name_=channel_name)
         with self.assertRaises(InvalidInputValueError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=0, channel_name_='bar')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=0, channel_name_=channel_name)
             
         # channel_name_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_=42)
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=42)
 
         # channel_limits_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', channel_limits_=42)
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, channel_limits_=42)
         with self.assertRaises(InvalidInputValueError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', channel_limits_=['-12.3', 13.5])
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, channel_limits_=['-12.3', 13.5])
             
         # channel_frequency_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', channel_frequency_='42')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, channel_frequency_='42')
         with self.assertRaises(InvalidInputValueError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', channel_frequency_=0)
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, channel_frequency_=0)
 
         # novelty_frequency_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', novelty_frequency_='42')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, novelty_frequency_='42')
         with self.assertRaises(InvalidInputValueError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', novelty_frequency_=0)
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, novelty_frequency_=0)
             
         # novelty_duration_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', novelty_duration_='42')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, novelty_duration_='42')
         with self.assertRaises(InvalidInputValueError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', novelty_duration_=-42)
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, novelty_duration_=-42)
             
         # novelty_impact_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', novelty_impact_='42')
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, novelty_impact_='42')
 
         # channel_name_
         with self.assertRaises(InvalidInputTypeError):
-            invalid_generator = man.create_pipeline(ip_='localhost', port_=1883, topic_='foo', frequency_=1, channel_name_='bar', pipeline_name_=42)
+            invalid_generator = man.create_pipeline(ip_=ip, port_=port, topic_=topic, frequency_=frequency, channel_name_=channel_name, pipeline_name_=42)